third_party/WebKit/LayoutTests/external/wpt/html/browsers/history/the-location-interface/location_hash_usvstring.html

<!DOCTYPE HTML>
<html>
  <head>
    <title>location_hash_usvstring</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
    test(function () {
      // USVString is equivalent to DOMString except for not allowing unpaired
      // surrogate codepoints. U+D800 to U+DFFF code point values are reserved
      // as high and low surrogates. Unpaired surrogate codepoints present in
      // USVString are replaced with U+FFFD.
      location.hash = '\uD999';
      strActual = location.hash;
      location.hash = '\uFFFD';
      strReplace = location.hash;
      assert_equals(strActual, strReplace);
    }, "unpaired surrogate codepoint should be replaced with U+FFFD");
    </script>
  </body>
</html>
